data-to-models/requirements.txt
###### Requirements without Version Specifiers ######
Pillow
regex
pandas
numpy
scikit-learn
tensorflow
keras
###### Requirements with Version Specifiers ######
